Move the matcher button back into the cog
